Refactored Search component into sub-components
